# docs/swagger.yaml
basePath: /
definitions:
  api.APIKeyResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
    type: object
...
  api.ConnectionResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
...
  api.CreateAPIKeyRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  api.CreateAPIKeyResponse:
    properties:
      api_key:
        type: string
      name:
        type: string
    type: object
  api.CreateConnectionRequest:
    properties:
      api_key:
        type: string
      name:
        type: string
      provider_id:
        type: string
    required:
    - api_key
    - name
    - provider_id
    type: object
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.ListAPIKeysResponse:
    properties:
      api_keys:
        items:
          $ref: '#/definitions/api.APIKeyResponse'
        type: array
    type: object
  api.ListConnectionsResponse:
    properties:
      connections:
        items:
          $ref: '#/definitions/api.ConnectionResponse'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Proxy server for Generative AI models.
  license:
    name: Apache 2.0
  title: GenAI Proxy API
  version: "1.0"
paths:
  /api/api-keys:
    get:
      consumes:
      - application/json
      description: List all API keys for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ListAPIKeysResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all API keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      description: Create a new API key for the authenticated user.
      parameters:
      - description: API key details
        in: body
        name: key
        required: true
        schema:
          $ref: '#/definitions/api.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.CreateAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new API key
      tags:
      - API Keys
  /api/api-keys/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an API key for the authenticated user.
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an API key
      tags:
      - API Keys
...
  /api/connections:
    get:
      consumes:
      - application/json
      description: List all connections for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ListConnectionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all connections
      tags:
      - Connections
    post:
      consumes:
      - application/json
      description: Create a new connection for the authenticated user.
      parameters:
      - description: Connection details
        in: body
        name: connection
        required: true
        schema:
          $ref: '#/definitions/api.CreateConnectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new connection
      tags:
      - Connections
  /api/connections/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a connection for the authenticated user.
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a connection
      tags:
      - Connections
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter your username and password to get a token.
    flow: password
    tokenUrl: /api/login
    type: oauth2
swagger: "2.0"